feat(web): complete input page ui
