+++ appeals/web/src/server/appeals/appeal-details/interested-party-comments/redact/components/form/buttons.js
@@ -1,0 +1,26 @@
+/** @type {PageComponent[]} */
+export const buttons = [
+	{
+		type: 'button',
+		parameters: {
+			text: 'Redact',
+			id: 'redact-button',
+			classes: 'govuk-button--secondary'
+		},
+		wrapperHtml: {
+			opening: '<div class="govuk-button-group">',
+			closing: '</div>'
+		}
+	},
+	{
+		type: 'button',
+		parameters: {
+			text: 'Confirm',
+			type: 'submit'
+		},
+		wrapperHtml: {
+			opening: '<div class="govuk-button-group">',
+			closing: '</div>'
+		}
+	}
+];
