+++ appeals/web/src/server/appeals/appeal-details/interested-party-comments/redact/components/instructions-list.js
@@ -1,0 +1,17 @@
+import { simpleHtmlComponent } from '#lib/mappers/components/html.js';
+
+const instructionsListMarkup = `
+<ol class="govuk-list govuk-list--number">
+  <li>
+    Use the ‘redacted comment’ box to select the words you want to redact.
+  </li>
+  <li>
+    Select ‘redact’ to confirm your redaction. Repeat until you have finished.  
+  </li>
+  <li>
+    Select ‘continue’ when you have finished redacting.
+  </li>
+</ol>
+`;
+
+export const instructionsList = simpleHtmlComponent(instructionsListMarkup);
