+++ appeals/web/src/server/appeals/appeal-details/interested-party-comments/redact/components/subtitle.js
@@ -1,0 +1,9 @@
+import { simpleHtmlComponent } from '#lib/mappers/components/html.js';
+
+const subtitleMarkup = `
+<h2>	
+  Make redactions
+</h2>
+`;
+
+export const subtitle = simpleHtmlComponent(subtitleMarkup);
